backend/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
pymysql
aiomysql
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
pypdf2
pdfplumber
python-docx
requests
zhipuai
faiss-cpu
pytest
pytest-asyncio